+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## [2023-04-09]
+
+- Order groups
+
 ## [2023-04-08]
 
 - Add _Contributing_ section in _README.md_.